# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TFSO API",
        "description": "API for TFSO application",
        "contact": {
            "name": "Burak",
            "url": "https://burakkaynar.com",
            "email": "burak@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Tüm kategorileri listeler",
                "operationId": "8e17a9627dbed245cf032f867569cf11",
                "responses": {
                    "200": {
                        "description": "Kategorilerin listesi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Sunucu hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kategoriler getirilirken bir hata oluştu."
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories/bulk": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Toplu kategori ekleme işlemi",
                "operationId": "9942246c1bc0e16e765b2d45e0b9dcb5",
                "requestBody": {
                    "description": "Eklenmek istenen kategorilerin listesi",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "categories": {
                                        "description": "Kategori objelerinin listesi",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "description": "Kategori adı",
                                                    "type": "string",
                                                    "example": "Pop Müzik"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kategoriler başarıyla eklendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kategoriler başarıyla eklendi."
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Doğrulama hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Doğrulama hatası"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "categories.0.name": [
                                                    "The categories.0.name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Sunucu hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kategoriler eklenirken bir hata oluştu."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "SQLSTATE[23000]: Integrity constraint violation..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories/{id}/tracks": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Bir kategoriye ait tüm şarkıları listeler",
                "operationId": "7607f5d38c720a8439fd49f8f639b8ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Kategori ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Şarkıların listesi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Track"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kategori bulunamadı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kategori bulunamadı."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
        "/api/categories/{id}/sets": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Bir kategoriye ait tüm setleri listeler",
                "operationId": "51bd8bd09378912d0e5e34426c563fc8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Kategori ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Setlerin listesi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Set"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kategori bulunamadı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kategori bulunamadı."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "title": "Category",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Pop"
                    },
                    "image": {
                        "type": "string",
                        "example": "categories/pop.jpg",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "http://localhost/storage/categories/pop.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
            "Set": {
                "title": "Set",
                "description": "Model representing a set of tracks",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the set",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true
                    },
                    "user_id": {
                        "description": "ID of the user who created the set",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Name of the set",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the set",
                        "type": "string"
                    },
                    "cover_image": {
                        "description": "URL of the cover image",
                        "type": "string"
                    },
                    "audio_file": {
                        "description": "Path to the audio file",
                        "type": "string"
                    },
                    "is_premium": {
                        "description": "Indicates if the set is premium",
                        "type": "boolean"
                    },
                    "iap_product_id": {
                        "description": "In-app purchase product ID for premium access",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp of the set",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Last update timestamp of the set",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "Track": {
                "title": "Track",
                "description": "Track model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "audio_file": {
                        "type": "string"
                    },
                    "image_file": {
                        "type": "string"
                    },
                    "is_premium": {
                        "type": "boolean"
                    },
                    "iap_product_id": {
                        "type": "string"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "audio_url": {
                        "type": "string",
                        "readOnly": "true"
                    },
                    "image_url": {
                        "type": "string",
                        "readOnly": "true"
                    },
                    "audio_file_name": {
                        "type": "string",
                        "readOnly": "true"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            }
        }
    },
...
}